% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtanb.R
\name{dtanb}
\alias{dtanb}
\title{The density function of the Tanget Burr Type XII probability distribution.}
\usage{
dtanb(x, c, k, s)
}
\arguments{
\item{x}{vector of quantiles.}

\item{c}{C parameter.}

\item{k}{K parameter.}

\item{s}{S parameter.}
}
\value{
A vector with n observations of the Tanget Burr Type XII distribution.
}
\description{
The density function of the Tanget Burr Type XII probability distribution.
}
\examples{
dtanb(0.5,32.5,3,3.5)
dtanb(0.5,2,3,3)
}

