% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtanb.R
\name{qtanb}
\alias{qtanb}
\title{The quantile function of the Tanget Burr Type XII probability distribution.}
\usage{
qtanb(p, c, k, s, lower = TRUE, log.p = FALSE)
}
\arguments{
\item{p}{Vector of probabilities.}

\item{c}{C parameter.}

\item{k}{K parameter.}

\item{s}{S parameter.}

\item{lower}{Lower parameter.}

\item{log.p}{Log.p parameter.}
}
\value{
A vector with n observations of the Tanget Burr Type XII distribution.
}
\description{
The quantile function of the Tanget Burr Type XII probability distribution.
}
\examples{
ptanb(0.5,1,1,1,TRUE,FALSE)
ptanb(0.5,2,1,1,TRUE,FALSE)
}

