% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanb.R
\name{stanb}
\alias{stanb}
\title{The survival function of the Tanget Burr Type XII probability distribution.}
\usage{
stanb(x, c, k, s)
}
\arguments{
\item{x}{vector of quantiles.}

\item{c}{C parameter.}

\item{k}{K parameter.}

\item{s}{S parameter.}
}
\value{
A vector with n observations of the Tanget Burr Type XII distribution.
}
\description{
The survival function of the Tanget Burr Type XII probability distribution.
}
\examples{
ptanb(0.5, 1, 1, 1)
ptanb(0.5, 2, 1, 1)
}

