% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TanB.R
\docType{package}
\name{TanB}
\alias{TanB}
\alias{TanB-package}
\alias{TanB-package}
\title{TanB: The Tangent-Skewed Burr Type XII distribution.}
\description{
The TanB package provides functions in the model of R's default
dnorm, rnorm and pnorm.
}
\details{
Details
}
\section{TanB functions}{

\code{\link{dtanb}}
\code{\link{htanb}}
\code{\link{ptanb}}
\code{\link{qtanb}}
\code{\link{rtanb}}
\code{\link{stanb}}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TrigonometricDistribution}
  \item Report bugs at \url{https://github.com/TrigonometricDistribution/TanB/issues}
}

}
